Further work on processing data through datastream filters
Previously we wrote filter output to a file and passed the path to
this file to the next filter. However, some filters will modify the
input data file in place, which means that we lose the output of
individual filters. To ensure that we can track back through the
output of individual filters if desired, pass the data via standard
input channel, so that modifications made by filter are applied to
a temporary file copy instead of our stored version.

Also, add some attributes to the Data class, add a method to
Datastream to log events, and add a new test program.