Import code from CUPS required to look up filter chain via mime types
Brings in CUPS code to provide methods for determining the filter
chain using mime.types and mime.convs.

The imported code is the minimum required to access the desired
methods, with heavy modification to avoid code that's not required (or
desired).

A test program allows us to look up the filter chain for specified
input and target destination.